factor out Bassetlaw mods into a base class
